# static/swagger.json
{"swagger": "2.0", "info": {"version": "1.0.0", "title": "Key-Value storage API"}, "paths": {"/getValues": {"get": {"summary": "Get all values", "description": "", "responses": {"200": {"description": "Set of key-value pairs"}, "404": {"description": "No values stored or wrong paginating"}}, "parameters": [{"in": "query", "name": "page", "required": false, "type": "integer", "minimum": 1, "default": 1, "description": "Number of the page"}, {"in": "query", "name": "per_page", "required": false, "type": "integer", "minimum": 1, "default": 10, "description": "How many items to be shown on the page"}], "tags": ["keyValue"]}}, "/getKeys": {"get": {"summary": "Get all keys", "description": "", "responses": {"200": {"description": "Set of key-value pairs"}, "404": {"description": "No keys stored or wrong paginating"}}, "parameters": [{"in": "query", "name": "page", "required": false, "type": "integer", "minimum": 1, "default": 1, "description": "Number of the page"}, {"in": "query", "name": "per_page", "required": false, "type": "integer", "minimum": 1, "default": 10, "description": "How many items to be shown on the page"}], "tags": ["keyValue"]}}, "/getAll": {"get": {"summary": "Get all key-value pairs", "description": "", "responses": {"200": {"description": "Set of key-value pairs"}, "404": {"description": "No pairs stored or wrong paginating"}}, "parameters": [{"in": "query", "name": "page", "required": false, "type": "integer", "minimum": 1, "default": 1, "description": "Number of the page"}, {"in": "query", "name": "per_page", "required": false, "type": "integer", "minimum": 1, "default": 10, "description": "How many items to be shown on the page"}], "tags": ["keyValue"]}}, "/clear": {"get": {"summary": "Truncate all values from the storage", "description": "", "responses": {"200": {"description": "Set of key-value pairs"}}, "tags": ["keyValue"]}}, "/set": {"get": {"summary": "Create new pair, update if it exists", "description": "", "responses": {"200": {"description": "Set of key-value pairs"}, "400": {"description": "Key or value are not set correctly"}}, "parameters": [{"in": "query", "name": "k", "required": true, "type": "string", "minimum": 1, "maximum": 64, "description": "The name of the key"}, {"in": "query", "name": "v", "required": true, "type": "string", "minimum": 1, "maximum": 256, "description": "The value for the given key"}], "tags": ["keyValue"]}}, "/get": {"get": {"summary": "Get pair by key", "description": "", "responses": {"200": {"description": "Set of key-value pairs"}, "400": {"description": "The key is not set correctly"}, "404": {"description": "No value saved for the key"}}, "parameters": [{"in": "query", "name": "k", "required": true, "type": "string", "minimum": 1, "maximum": 64, "description": "The name of the key"}], "tags": ["keyValue"]}}, "/is": {"get": {"summary": "Check if pair exists", "description": "", "responses": {"200": {"description": "Return True if pair exists"}, "400": {"description": "The key is not set correctly"}, "404": {"description": "Return False if pair doesn`t exists"}}, "parameters": [{"in": "query", "name": "k", "required": true, "type": "string", "minimum": 1, "maximum": 64, "description": "The name of the key"}], "tags": ["keyValue"]}}, "/rm": {"get": {"summary": "Remove a pair", "description": "", "responses": {"200": {"description": "Set of key-value pairs"}, "400": {"description": "The key is not set correctly"}, "404": {"description": "No value stored for that key"}}, "parameters": [{"in": "query", "name": "k", "required": true, "type": "string", "minimum": 1, "maximum": 64, "description": "The name of the key"}], "tags": ["keyValue"]}}}, "definitions": {}, "servers": {"url": ["/"]}}